==> Notebooks/1.0 broze- pontos de interessev02.ipynb <==
!pip install googlemaps

import googlemaps
import pandas as pd
import requests

# define a chave de acesso da API
api_key = 'AIzaSyBKFp_hemLkS16tEY8Yp7Xa7JEOG_Si_3o'

# cria o objeto do cliente para a API do Google Maps
gmaps = googlemaps.Client(api_key)

# define a latitude e longitude da região de interesse
latitude = -23.5505
longitude = -46.6333

# define o raio máximo em metros para considerar um supermercado na região
radius = 1000

# faz a consulta na API para encontrar os supermercados na região
results = gmaps.places_nearby(location=(latitude, longitude), radius=radius, type='supermarket')

def count_supermarket_lat_long (lat, lon, radius=1000):
    
    # define o raio máximo em metros para considerar um supermercado na região
    # radius = 1000
    
    results = gmaps.places_nearby(location=(lat, lon), radius=radius, type= ["pharmacy","supermarket", "tourist_attraction", "restaurant", "subway_station"])
    count_supermarket = len(results['results'])
    
    return count_supermarket
    



def count_supermarket_lat_long (lat, lon, radius=1000):
    
    # define o raio máximo em metros para considerar um supermercado na região
    # radius = 1000
    
    results = gmaps.places_nearby(location=(lat, lon), radius=radius, type= ["pharmacy","supermarket", "tourist_attraction", "restaurant", "subway_station"])
    count_supermarket = len(results['results'])
    
    return count_supermarket

count_supermarket_lat_long (-23.5505,-46.6333)

results.keys(),results['results'],[{},{},{}]

mount_path = "/workspaces/prediction_house_price_from_airbnb/Data"

df = pd.read_parquet(mount_path + "/bronze/listings.parquet")

dados=df.iloc[0:10][['id','latitude','longitude']]

dados


df['qtde_supermaket'] = df.apply(lambda row: count_supermarket_lat_long(row['latitude'], row['longitude']), axis=1)

df

# Importar as bibliotecas necessárias
# import requests
# import pandas as pd

# Função para obter os pontos de interesse dentro de um raio usando a API do Google Places
def obter_pontos_interesse(latitude, longitude, raio, tipos_lugar, api_key):
    url = 'https://maps.googleapis.com/maps/api/place/nearbysearch/json'
    params = {
        'location': f'{latitude},{longitude}',
        'radius': raio,
        'types': '|'.join(tipos_lugar),
        'key': api_key
    }
    response = requests.get(url, params=params)
    data = response.json()

    return data['results']


# Raio em metros (convertido para o formato exigido pela API)
raio = 1000

# Tipos de lugar para filtrar os pontos de interesse
tipos_lugar = {
    'Pharmacy': 'pharmacy',
    'Grocery': 'grocery_or_supermarket',
    'Tourist Attraction': 'tourist_attraction',
    'Restaurant': 'restaurant',
    'Subway Station': 'subway_station'
}

# Chave da API do Google Places
# api_key = 'sua_chave_da_api_aqui'

# Iterar sobre as linhas do dataframe
for index, row in df.iterrows():
    latitude = row['latitude']
    longitude = row['longitude']
    
    # Iterar sobre os tipos de lugar
    for tipo, tipo_lugar in tipos_lugar.items():
        # Obter os pontos de interesse dentro do raio e tipo de lugar usando a API do Google Places
        resultado = obter_pontos_interesse(latitude, longitude, raio, [tipo_lugar], api_key)
        
        # Adicionar coluna para o tipo de lugar, com valor 1 se houver pontos de interesse desse tipo, 0 caso contrário
        df.loc[index, tipo] = len(resultado) if resultado else 0

# Preencher NaN com 0 nas colunas de tipos de lugar
df.fillna(0, inplace=True)

# Exibir as primeiras linhas do dataframe resultante
df.head()
