==> Notebooks/2.0_silver_pontos.ipynb <==
import pandas as pd


path = '/workspaces/prediction_house_price_from_airbnb/Data/'

df01=pd.read_parquet(path=path +'bronze/pontos_interesses_complete.parquet')

df=pd.read_parquet(path=path+'silver/listings_v01.parquet')

df=df.merge(df01, on='id')

df.to_parquet(path=path+'gold/listings_pontos.parquet')
