==> Notebooks/3.0_prediction-01.ipynb <==
import pandas as pd 
import numpy as np
from sklearn.linear_model import LinearRegression
from sklearn.feature_selection import SelectKBest, chi2
from sklearn.model_selection import train_test_split

mount_path = "/workspaces/prediction_house_price_from_airbnb/Data"

df=pd.read_parquet(path=mount_path+'/gold/listings_pontos.parquet')

pd.set_option('display.max_columns', None)

column_names = list(df.columns)
column_names

# Selecione as características que deseja usar para prever o preço

features = ['id',
 'source',
 'name',
 'description',
 'neighborhood_overview',
 'neighbourhood',
 'neighbourhood_cleansed',
 'property_type',
 'room_type',
 'accommodates',
 'bathrooms',
 'bathrooms_text',
 'bedrooms',
 'beds',
 'amenities',
 'minimum_nights',
 'maximum_nights',
 'minimum_minimum_nights',
 'maximum_minimum_nights',
 'minimum_maximum_nights',
 'maximum_maximum_nights',
 'minimum_nights_avg_ntm',
 'maximum_nights_avg_ntm',
 'calendar_updated',
 'has_availability',
 'availability_30',
 'availability_60',
 'availability_90',
 'availability_365',
 'calendar_last_scraped',
 'number_of_reviews',
 'number_of_reviews_ltm',
 'number_of_reviews_l30d',
 'first_review',
 'last_review',
 'review_scores_rating',
 'review_scores_accuracy',
 'review_scores_cleanliness',
 'review_scores_checkin',
 'review_scores_communication',
 'review_scores_location',
 'review_scores_value',
 'license',
 'instant_bookable',
 'calculated_host_listings_count',
 'calculated_host_listings_count_entire_homes',
 'calculated_host_listings_count_private_rooms',
 'calculated_host_listings_count_shared_rooms',
 'reviews_per_month',
 'distance_centro',
 'distance_corcovado',
 'distance_galeao',
 'distance_olipico',
 'distance_copacabana',
 '9',
 'T',
 'Z',
 '0',
 'R',
 '5',
 'y',
 'r',
 'Y',
 'i',
 'v',
 '/',
 '2',
 'S',
 ']',
 '$',
 '+',
 'q',
 ',',
 ':',
 'V',
 't',
 's',
 'z',
 '6',
 "'",
 'f',
 'K',
 'n',
 'U',
 '[',
 'H',
 '3',
 'b',
 'F',
 ';',
 '.',
 'g',
 'L',
 'O',
 'h',
 'M',
 'E',
 '7',
 'B',
 'D',
 'd',
 'a',
 '*',
 'm',
 'G',
 'X',
 'W',
 'P',
 '\\',
 'c',
 '&',
 'k',
 'o',
 'x',
 'e',
 '%',
 '`',
 'A',
 'l',
 '|',
 '4',
 '1',
 '(',
 'N',
 'w',
 ')',
 'I',
 '-',
 '!',
 'J',
 'u',
 '8',
 'j',
 ' ',
 'C',
 'p',
 'Q',
 'years',
 'Pharmacy',
 'Grocery',
 'Tourist Attraction',
 'Restaurant',
 'Subway Station']

# Divida o conjunto de dados em um conjunto de treinamento e teste
X_train, X_test, y_train, y_test = train_test_split(df[features], df["price"], test_size=0.25)



# Selecione as 5 características mais importantes
selector = SelectKBest(score_func=chi2, k=5)
X_train_selected = selector.fit_transform(X_train, y_train)
X_test_selected = selector.transform(X_test)

# Crie um modelo de regressão linear
model = LinearRegression()



# Treine o modelo
model.fit(X_train_selected, y_train)

# Preveja os preços das casas no conjunto de teste
y_pred = model.predict(X_test_selected)



# Avalie a precisão do modelo
print("Accuracy:", model.score(X_test_selected, y_test))
